+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### adsb_deku
 
 ### apps/radar
-- [radar] fix breaking clap change, same syntax as before for `--cities`.
+- fix breaking clap change, same syntax as before for `--cities`.
+- fix time related unwrap(). Thanks ([@Jachdich](https://github.com/Jachdich)) ([!57](https://github.com/rsadsb/adsb_deku/pull/57))
 
 ### apps/1090