+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### adsb_deku
 - refactor: improve `fmt::Display` of `ControlField`
+- fix(breaking): change `OperationalCodeSurface.reserved` from `u16` to `u8`
 
 ### apps/radar
 - fix: breaking clap change, same syntax as before for `--cities`.