+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: add version info to logging
 - feat: improve user facing errors
 - feat(breaking): use mercator projection for map/coverage tabs, change `--scale` usage
+- feat: Update to clap v3.0.0
 
 ### apps/1090